+++ lib/src/code_generator/dart_keywords.dart
@@ -66,5 +66,6 @@ const keywords = {
   'show',
   'dynamic',
   'implements',
-  'static'
+  'static',
+  'late',
 };