+++ lib/src/config_provider/config_types.dart
@@ -26,11 +26,13 @@ class CommentType {
 enum CommentStyle { doxygen, any }
 enum CommentLength { none, brief, full }
 
+enum StructDependencies { full, opaque }
+
 /// Represents a single specification in configurations.
 ///
 /// [E] is the return type of the extractedResult.
 class Specification<E> {
-  final bool Function(String name, dynamic value) validator;
+  final bool Function(List<String> name, dynamic value) validator;
   final E Function(dynamic map) extractor;
   final E Function()? defaultValue;